+++ src/runtime/gc_globals_conservative.go
@@ -1,5 +1,5 @@
-//go:build (gc.conservative || gc.extalloc) && (baremetal || tinygo.wasm)
-// +build gc.conservative gc.extalloc
+//go:build gc.conservative && (baremetal || tinygo.wasm)
+// +build gc.conservative
 // +build baremetal tinygo.wasm
 
 package runtime
