+++ src/runtime/gc_stack_portable.go
@@ -1,6 +1,5 @@
-//go:build (gc.conservative || gc.extalloc) && tinygo.wasm
-// +build gc.conservative gc.extalloc
-// +build tinygo.wasm
+//go:build gc.conservative && tinygo.wasm
+// +build gc.conservative,tinygo.wasm
 
 package runtime
