+++ src/runtime/gc_stack_raw.go
@@ -1,5 +1,5 @@
-// +build gc.conservative gc.extalloc
-// +build !tinygo.wasm
+//go:build gc.conservative && !tinygo.wasm
+// +build gc.conservative,!tinygo.wasm
 
 package runtime
