--- src/runtime/runtime_unix_noheap.go
@@ -1,9 +1,0 @@
-// +build darwin linux,!baremetal,!wasi
-
-// +build !nintendoswitch
-
-// +build gc.none gc.extalloc
-
-package runtime
-
-func preinit() {}
